requirements.txt
pnumpy
torch
torchvision
pretty_midi
pypianoroll
mido
pandas
scipy
matplotlib
jupyter
tqdm
scikit-learn